content/media/test/test_wav_standalone.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Media test: standalone video documents</title>
  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body onload="doTest()">

<iframe id="i" src="r11025_u8_c1.wav"></iframe>

<pre id="test">
<script class="testbody" type="text/javascript">
SimpleTest.waitForExplicitFinish();
var gExpectEnded = 2;

function endTest()
{
  var elem = document.getElementById("i").contentDocument.body.firstChild;
  gExpectEnded--;
  if (gExpectEnded == 1) {
    elem.play();
  } else {
    is(gExpectEnded, 0, "not expecting more ended events");
    SimpleTest.finish();
  }
}

function doTest()
{
  var elem = document.getElementById("i").contentDocument.body.firstChild;
  is(elem.tagName.toLowerCase(), "video", "Is video element");
  is(elem.currentSrc.substring(elem.currentSrc.length - 16), "r11025_u8_c1.wav", "currentSrc");
  is(elem.controls, true, "Controls set");
  is(elem.autoplay, true, "Autoplay set");

  if (elem.ended) {
    gExpectEnded = 1;
    elem.play();
  }
  elem.addEventListener("ended", endTest, false);
}
</script>
</pre>
</body>
</html>
